requirements/base.txt
# requirements/base.txt
mss>=9.0.0
opencv-python>=4.8.0
websockets>=12.0
pystray>=0.19.0
dearpygui>=1.11.0
pynput>=1.7.6
Pillow>=10.0.0
requests>=2.31.0
numpy>=1.26.0
pytesseract>=0.3.10
psutil>=5.9.0
